requirements.txt
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
mysql-connector-python==8.3.0
pdfminer==20191125
pdfminer.six==20231228
pycparser==2.22
pycryptodome==3.20.0
